% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_message.R
\name{add_http_status_guidance}
\alias{add_http_status_guidance}
\title{Adds Extracted HTTP Status Guidance to Original API Error Message}
\usage{
add_http_status_guidance(e, api_error)
}
\arguments{
\item{e}{Error condition. An HTTP request error condition to extract the status code from. Required.}

\item{api_error}{Character. The original API/client error message. Required.}
}
\value{
This function returns a character string that contains the original API error message and the HTTP status code-specific guidance extracted by \code{\link{http_status_guidance}} as a character string, separated by a line.
}
\description{
This function appends the suggested guidance for a given HTTP status code, extracted by \code{\link{http_status_guidance}} to the original error message associated with the API service. Both are presented in the ChatRBox dynamic error message.
}
\details{
This function combines the helper functions \code{\link{http_status_from_condition}} and \code{\link{http_status_guidance}} to determine the user-facing error message associated with HTTP status errors. The dynamic error message for any failed \code{do.call()} contains maximum information for user trouble-shooting, including chatbot-relevant parameters (chosen client paths and parameters), request-relevant parameters (\pkg{httr2} request assignments) and API-relevant parameters (HTTP status codes). This function is concerned with the API-relevant parameters and ensures that the original, simple, API error is presented alongside ChatRBox-specific advice for any HTTP status.
}
\examples{
# A dynamic error message outputs for any failed AI-facilitated do.call()
# This contains attempted client path and parameters, API HTTP error, httr2 
# configuration used by the client function and suggested next steps  

# Example HTTP error status
error_condition <- structure(
  list(status_code = 414),
  class = "error"
)

# Extracts error code to assign ChatRBox-specific guidance
status_code <- http_status_from_condition(error_condition)
status_code

# Assigns ChatRBox guidance to extracted code
guidance <- http_status_guidance(status_code)

# Example external API error message
api_error <- "URL too long"

# Final error message depicts both general API warning and assigned 
# ChatRBox guidance
error <- add_http_status_guidance(error_condition, api_error) 
}
\seealso{
\code{\link{http_status_from_condition}}, \code{\link{http_status_guidance}}, \code{\link{llm_api_result}}
}
