% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_token.R
\name{check_names}
\alias{check_names}
\title{Ensures Argument is a Named List or Vector}
\usage{
check_names(arg, arg_label, input_label)
}
\arguments{
\item{arg}{List or vector. The object to check for names. Required.}

\item{arg_label}{Character. The label (argument name) to display in error messages for \code{arg}. Required.}

\item{input_label}{Character. The label (argument name) to display in error messages for the expected input target. Required.}
}
\value{
This function is called for its side effect and will throw an error if \code{arg} is not a named list or vector.
}
\description{
Checks that the provided argument is named. Throws an error if any elements are not named.
}
\details{
This function is an internal checker on the validity of inputted API tokens. Tokens and token headers are assigned to the corresponding API service URL in \code{\link{get_api_client_env}} or OpenAPI JSON schema URL in \code{\link{build_api_client_env}}, based on name matching between named lists. Token and header names must be present, identical to an inputted URL and unique. This function checks that they are present. The remaining checks are conducted using \code{\link{validate_name_match}} and \code{\link{check_duplicate_names}}. Similar name matching logic is used to assign summary prompts to given API services, within \code{\link{process_summary_list}}. Summary logic is documented in \code{chatrbox_ai_services.R}, as this workflow is triggered when \code{summarize} is set to \code{TRUE} within ChatRBox \code{talk} calls.
}
\examples{
# Lists to compare represent service and token/summary prompt names
arg <- c(alpha = 1, beta = 2)
main <- c(alpha = 10, beta = 20, gamma = 30)

# arg is a named list, leading to no error and NULL result
result <- check_names(
  arg = arg,
  arg_label = "arg",
  input_label = "main"
)
result

# alpha and beta are present in main, leading to no error and NULL result
result <- validate_name_match(
  name_list = arg,
  main_list = main,
  arg_label = "name_list",
  target_label = "main_list"
)
result

# arg has unique names, leading to no error and NULL result 
check_duplicate_names(
  arg = arg,
  arg_label = "arg"
)

# These helpers were defined for their side effects, meaning this example code
# will only lead to NULL assignments. Failing tests will error with a
# descriptive message
}
\seealso{
\code{\link{get_api_client_env}}, \code{\link{build_api_client_env}}, \code{\link{validate_name_match}}, \code{\link{check_duplicate_names}}
}
