% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ai_response.R
\name{code_extract}
\alias{code_extract}
\title{Extracts Code Block from Markdown String}
\usage{
code_extract(string = NULL, language = "json")
}
\arguments{
\item{string}{Character. The input string containing Markdown code block(s). Required.}

\item{language}{Character. The language tag of the code block. One of the supported tags in \code{.SUPPORTED_CODE_LANGUAGES} (\code{"json"}, \code{"display"}, \code{"yaml"}, \code{"yml"}, \code{"tsv"}, \code{"csv"}, \code{"md"}). This may be extended by adding to the vector assignment. See CONTRIBUTING.md for further detail. Defaults to "json".}
}
\value{
The contents of the first extracted code block.
}
\description{
This function extracts the first code block of a specified language from a Markdown-formatted string. The extracted block excludes language tags and code fences.
}
\details{
This function is used in \code{\link{llm_api_result}} to parse the LLM response. The ChatRBox AI response given the default prompt should contain both a JSON object and text output, if an API service/tool has been used. The JSON object is used to populate the chosen API service/tool function with parameters from the input question, whereas the explanation is for user interaction. Hence, these are employed differently in downstream workflows. 

This function separates the JSON object and outputs the contained client function arguments. The equivalent function for obtaining the text output is \code{\link{code_remove}}.
}
\examples{
# Example AI response using linear plotting API, 
# consisting of dialogue and code sections
response <- '```json
{ 
  "path": "services$api_services$plot$client_plot",
  "name": "linearplotting",
  "df": "example_data",
  "x_col": "",
  "y_col": "",
  "title": "",
  "api_url": "http://127.0.0.1:8000/plot",
  "token": null
}
```
The linear plotting API allows for the visualization of 
bivariate data sets, summarized by a line of best fit!'

# Extracts dialogue 
dialogue <- code_remove(string = response, language = "json")
dialogue 

# Extracts code
code <- code_extract(string = response, language = "json")
code

# Retrieves name key value
name <- get_name(json_str = code)
name

# Retrieves latter two components of path string
# This is used to create dynamic error message for failed do.call()
path <- extract_path("services$api_services$plot$client_plot")
path
}
\seealso{
\code{\link{llm_api_result}}, \code{\link{code_remove}}
}
