% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_api_client.R
\name{decode_api_response}
\alias{decode_api_response}
\title{Decodes API Responses into R Objects}
\usage{
decode_api_response(resp, simplify_json = TRUE)
}
\arguments{
\item{resp}{Response body. \pkg{httr2} response object containing content type. Required.}

\item{simplify_json}{Logical. Passed to \code{jsonlite::fromJSON(simplifyVector = ...)}. Defaults to \code{TRUE}.}
}
\value{
Decoded response object.
}
\description{
This function decodes an \pkg{httr2} response according to its declared content type. This function does not render, reshape, serialize, coerce, or otherwise assume what downstream APIs require.
}
\details{
JSON is parsed into R objects. By default, \code{simplify_json = TRUE} preserves \pkg{jsonlite} default method, whereby homogeneous JSON arrays may become data frames. This functionality may be altered via modification of source code. This function is used within \code{\link{build_api_client}} as the final step of every generated client function, decoding the \pkg{httr2} response returned by \code{httr2::req_perform} according to its declared content type.
}
\seealso{
\code{\link{build_api_client}}
}
