% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ai_response.R
\name{extract_path}
\alias{extract_path}
\title{Extracts Latter Two Components of Function Path}
\usage{
extract_path(path = NULL)
}
\arguments{
\item{path}{Character. The '$'-delimited string representing a function path. Required.}
}
\value{
A named list with elements \code{client} and \code{service}, containing the extracted components.
}
\description{
This function extracts the last two names from a '$'-delimited string, representing the path to a client function. Therefore, the last component represents the function whereas the penultimate component represents the environment.
}
\details{
This function returns a list with these components named \code{client} and \code{service}, respectively. If fewer than two components are provided, both elements will be \code{NA}.


This function is used in \code{\link{llm_api_result}} when populating the chosen API service/tool with the extracted arguments fails. By locating the client function and environment chosen by the AI, a dynamic error message is outputted to the user. This informs them of an unsuccessful LLM response to an input question given a failed \code{do.call()} in approachable language. It is thought that users may more easily understand the attempted client and environment of the failed call than the full client function path. Hence, these have been extracted such that the user question, prompt, services or data may be refined.
}
\examples{
# Example AI response using linear plotting API, 
# consisting of dialogue and code sections
response <- '```json
{ 
  "path": "services$api_services$plot$client_plot",
  "name": "linearplotting",
  "df": "example_data",
  "x_col": "",
  "y_col": "",
  "title": "",
  "api_url": "http://127.0.0.1:8000/plot",
  "token": null
}
```
The linear plotting API allows for the visualization of 
bivariate data sets, summarized by a line of best fit!'

# Extracts dialogue 
dialogue <- code_remove(string = response, language = "json")
dialogue 

# Extracts code
code <- code_extract(string = response, language = "json")
code

# Retrieves name key value
name <- get_name(json_str = code)
name

# Retrieves latter two components of path string
# This is used to create dynamic error message for failed do.call()
path <- extract_path("services$api_services$plot$client_plot")
path
}
\seealso{
\code{\link{llm_api_result}}
}
