% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_services.R
\name{get_api_client_env}
\alias{get_api_client_env}
\title{Generates Client Function Environments for a List of API Service URLs}
\usage{
get_api_client_env(
  services_list,
  token_service = list(),
  header_service = list(),
  ...
)
}
\arguments{
\item{services_list}{List. Named list of API service base URLs as character strings, without trailing slashes. Required.}

\item{token_service}{List. Named list of API authentication tokens, with names matching APIs in \code{services_list}. Each token is used only for the matching service via pair matching. Error messages will throw for mismatched, duplicated and missing names. Defaults to empty list.}

\item{header_service}{List. Named list of authorization header names for each token. Each header is used only for the matching service via pair matching. Error messages will throw for mismatched, duplicated and missing names. Defaults to empty list. Users may dictate authorization method using this argument within the \code{\link{ChatRBox}} R6 object or edit the source code in \code{\link{get_api_client_fns}} which includes the \code{auth_scheme} argument.}

\item{...}{Additional arguments passed to \code{\link{get_api_client_fns}}.}
}
\value{
A named list of environments, one for each service in \code{services_list}, where each environment contains the generated client functions for that service. Names correspond to those inputted in \code{services_list}. This nested environment structure prevents namespace conflict, especially for multiple external API services.
}
\description{
This function iterates over a named list of API service URLs, extracting R client functions for each from the \code{client_fns} API endpoint using \code{\link{get_api_client_fns}}. These are returned as a named list of environments. Authentication tokens and custom headers are assigned to specific services through named lists, if required to access protected APIs.
}
\details{
This function applies \code{\link{get_api_client_fns}} to every API service URL in \code{services_list} in order to extract the client functions for each endpoint, for chatbot access. It is assumed that these API services have the endpoint \code{client_fns} containing client functions as R source code. These functions also enable the use of tokens and token headers to access protected APIs, whereby tokens may be entered as character strings or environment variables. This function and \code{\link{get_api_client_fns}} have equivalent functions called \code{\link{build_api_client_env}} and \code{\link{build_api_client}} for when the \code{client_fns} endpoint is not present. These rather generate client function R source code for each endpoint in an OpenAPI JSON schema and can also be utilized with access tokens. Notably, separate token and token header lists are required for APIs loaded as service URLs or OpenAPI JSON schemas, and schemas must be provided via the \code{openapi_list} rather than \code{services_list} argument upon chatbot initialization.
}
\examples{
# Simulates deployment of remote API service using the linear plotting API, 
# stored locally as plumber.R in inst/api_service 
proc <- start_api_service()

# Allows time for API service to launch in background R process
Sys.sleep(2) 

# Returns TRUE if process is created
proc$is_alive()

# Retrieves client functions from API URL (default host/port)
\dontrun{
  client_envir <- get_api_client_fns("http://127.0.0.1:8000")
  ls(client_envir)
}
}
\seealso{
\code{\link{get_api_client_fns}}, \code{\link{build_api_client_env}}, \code{\link{build_api_client}}
}
