% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_services.R
\name{get_api_client_fns}
\alias{get_api_client_fns}
\title{Fetch and Install R Client Functions from a Remote API}
\usage{
get_api_client_fns(
  base_url,
  path = "client_fns",
  token = Sys.getenv("ChatRBox_TOKEN", NA),
  auth_scheme = NULL,
  token_header = "X-API-Key",
  timeout = 60,
  method = "GET",
  verbose = FALSE,
  client_envir = new.env()
)
}
\arguments{
\item{base_url}{Character. The root URL of the API (without trailing slash). Required.}

\item{path}{Character. The path to the client code endpoint. Defaults to \code{"client_fns"}.}

\item{token}{Character. Optional API authentication token. Defaults to the value of \code{Sys.getenv("ChatRBox_TOKEN")}), retrieved from the global environment. This default is also true for the \code{token} argument in \code{\link{build_api_client}} (which is the equivalent to this function for OpenAPI JSON schemas rather than base URLs) since these are the helpers that extract client functions for a single API input. These functions are later iterated over input lists so that users may upload multiple API services simultaneously.}

\item{auth_scheme}{Character. API authorization scheme. Defaults to NULL, given \code{token_header} is populated. The user-facing R6 object \code{\link{ChatRBox}} only contains the \code{token_header} argument but users may choose authorization methods via \code{auth_scheme} by modifying this function.}

\item{token_header}{Character. API token header name if \code{auth_scheme} is \code{NULL}. Defaults to "X-API-Key".}

\item{timeout}{Numeric. Timeout value. Defaults to 60 seconds.}

\item{method}{Character. API HTTP method. Defaults to "GET" for likely chatbot API services.}

\item{verbose}{Logical. Whether to print verbose information. Defaults to \code{FALSE}.}

\item{client_envir}{Environment. The environment in which to evaluate (and store) the loaded client functions. Defaults to \code{new.env()}.}
}
\value{
The environment (\code{client_envir}) containing the evaluated client functions.
}
\description{
This function downloads R client source code from the \code{client_fns} endpoint of remote API services (typically Plumber servers). The downloaded code is parsed and evaluated into the specified environment, making client helper functions immediately available in R. The \code{client_fns} endpoint should serve plain R code, with an appropriate `Content-Type` header (e.g., `text/plain`). Optionally, if the endpoint serves JSON, this function will attempt to extract the code from the \code{$code} field. If an API key is required, it will be sent as a bearer token.
}
\details{
This function uses the helper function \code{\link{make_req}} and acts as a helper function to \code{\link{get_api_client_env}}. This function extracts client function R source code from the \code{client_fns} endpoint of an API service. These are downloaded such that they are available for user and chatbot use in R. Therefore, this function is specific to API services provided as API URLs via \code{services_list}, as it is assumed that these contain client function code, compared to those provided as OpenAPI JSON schema URLs via \code{openapi_list} which are not assumed to contain the \code{client_fns} endpoint. When APIs are provided via \code{openapi_list} instead, client functions are automatically generated by ChatRBox and collated with the client functions extracted from \code{services_list} APIs. Hence, chatbots view all API services equally and choose the most appropriate service path to fulfil user requests from the collated list.
}
\examples{
# Simulates deployment of remote API service using the linear plotting API, 
# stored locally as plumber.R in inst/api_service 
proc <- start_api_service()

# Allows time for API service to launch in background R process
Sys.sleep(2) 

# Returns TRUE if process is created
proc$is_alive()

# Retrieves client functions from API URL (default host/port)
\dontrun{
  client_envir <- get_api_client_fns("http://127.0.0.1:8000")
  ls(client_envir)
}
}
\seealso{
\code{\link{make_req}}, \code{\link{get_api_client_env}}
}
