% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_methods.R
\name{get_env_headers}
\alias{get_env_headers}
\title{Loads API Headers from a Serialized Environment Variable}
\usage{
get_env_headers(var = "API_HEADERS")
}
\arguments{
\item{var}{Character. Name of the environment variable holding JSON headers. Defaults to "API_HEADERS".}
}
\value{
A named character vector suitable for the \code{api_headers} argument of the \pkg{ellmer} function \code{chat_openai_compatible()}.
}
\description{
De-serializes JSON strings for use as named character vector arguments during chatbot initialization.
}
\details{
ChatRBox is compatible with any \pkg{ellmer} AI provider but initializing the \code{chat_ollama()} and \code{chat_openai_compatible()} functions have been documented. The latter contains the \code{api_headers} argument for extra headers appended to each API call. This must be a named character vector. Hence, users are advised to store this vector in an environment file as a single serialized (JSON) string. Loading this file ensures this environment variable is present within user R sessions, and calling this function de-serializes the string into the named vector during chatbot initialization.
}
