% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_methods.R
\name{get_function_path}
\alias{get_function_path}
\title{Constructs Environment/List Paths to Client Functions}
\usage{
get_function_path(obj, prefix = "")
}
\arguments{
\item{obj}{List. A list of environments/lists to search for functions. This may be nested. Required.}

\item{prefix}{Character. String used to build function paths during recursion. Defaults to empty string.}
}
\value{
A character vector of function paths beginning with \code{services}, using '$' notation.
}
\description{
This function recursively searches a list of environments/lists and outputs the paths to every function using '$' notation. Paths are outputted as character vectors.
}
\details{
This function is used within \code{\link{ChatRBox}} such that the top-level object of each path is hard-coded as \code{services}.


The output of this function documents the path to every function contained within a list/nested list. This function is applied to the list of services in \code{\link{ChatRBox}} to output a list of tool functions/API services paths. 

This is included as prompting in the AI request. Therefore, the LLM outputs the single most suitable function path to answer the input question. The equivalent function for function path extraction with arguments is \code{\link{get_function_args}}. Both listed outputs are included in the AI prompt.
}
\examples{
# Defines tool functions
add <- function(x,y) {
  x + y
}

subtract <- function(x,y) {
  x - y
}

# Constructs named tools list
tools_list = list(add_two_numbers = add,
                  subtract_two_numbers = subtract)

# Defines new tools environment
tools_env <- new.env()

# Adds tools list to tools environments
list2env(tools_list, envir = tools_env)

# Extracts tool function paths from tools environment
# services$ prefix is hard-coded
paths <- get_function_path(obj = tools_env)
paths

# Extracts arguments from tool functions 
function_args <- get_args_str(add)
function_args

# get_args_str() is used as a helper function to combine 
# function paths with arguments
# Extracts tool function paths with arguments from tools environment
args <- get_function_args(obj = tools_env)
args

# This process is used to construct function paths with arguments 
# for every API service/tool function. These are used in the default AI prompt 
}
\seealso{
\code{\link{ChatRBox}}
}
