% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_methods.R
\name{get_keyring}
\alias{get_keyring}
\title{Retrieves Values From Keyring}
\usage{
get_keyring(var = NULL, service = "ChatRBox")
}
\arguments{
\item{var}{Variable name. Required.}

\item{service}{Keyring service. Defaults to "ChatRBox", in accordance with \code{\link{store_keyring}} default values.}
}
\value{
The retrieved value as a string.
}
\description{
This function retrieves values from keyring using the environment variable name and service name. This is used to load environment variables (such as \pkg{ellmer} base URLs) without the use of a .Renviron file.
}
\details{
This function works with \code{\link{store_keyring}} to load stored user credentials within \code{\link{start_api_service}}, consistent with \pkg{callr} recommendations. This function uses the \pkg{callr} function \code{r_bg} to locally start the example API service, but temporarily makes a copy of user .Renviron files during this background process. Users are advised to store environmental variables in a YAML file named 'ChatRBox_config.yml', located in the home directory. The \code{examples.R} script in man/examples and the ChatRBox vignette include potential structures for this configuration file, with equivalent functionality for nested and non-nested variable retrieval. 

If users do not contain sensitive information in their .Renviron files or do not locally run the example API service using \code{\link{start_api_service}}, the \pkg{callr} background process poses no threat. Note that user .Renviron files are only copied into the R session's temporary directory, such that \code{callr::r_bg()} is safe and adheres to file operation guidance.
}
\examples{
# Store config values with keyring, retrieve during initialization
\dontrun{
store_keyring(config_path = "~/ChatRBox_config.yml")
  chat <- ellmer::chat_ollama(base_url = get_keyring("OLLAMA_HOST"),
                              model    = get_keyring("OLLAMA_MODEL"))
}
}
