% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ai_response.R
\name{get_name}
\alias{get_name}
\title{Returns Second Value from JSON Object}
\usage{
get_name(json_str = "")
}
\arguments{
\item{json_str}{Character. The input JSON object as a string. Defaults to empty string leading to \code{NA} output.}
}
\value{
The value corresponding to the second key-value pair in the inputted
JSON object.
}
\description{
This function returns the value from the second key-value pair in a JSON object, thought to correspond to API/tool service names for ChatRBox name-matching workflows. These are involved in applying AI-generated summary prompts and token assignments to the correct services.
}
\details{
The default prompt instructs the AI to assign a human-readable name to the value of the second key-value pair in its JSON output. The key for this pair is always \code{name}. Retrieving the value via positioning rather than referring to \code{name} was preferred, in case later arguments are also called \code{name}. This function is used when adding past API service/tool function outputs to the object environment to inform further AI prompting. Therefore, the chatbot can access API results to summarize or manipulate them further.
}
\examples{
# Example AI response using linear plotting API, 
# consisting of dialogue and code sections
response <- '```json
{ 
  "path": "services$api_services$plot$client_plot",
  "name": "linearplotting",
  "df": "example_data",
  "x_col": "",
  "y_col": "",
  "title": "",
  "api_url": "http://127.0.0.1:8000/plot",
  "token": null
}
```
The linear plotting API allows for the visualization of 
bivariate data sets, summarized by a line of best fit!'

# Extracts dialogue 
dialogue <- code_remove(string = response, language = "json")
dialogue 

# Extracts code
code <- code_extract(string = response, language = "json")
code

# Retrieves name key value
name <- get_name(json_str = code)
name

# Retrieves latter two components of path string
# This is used to create dynamic error message for failed do.call()
path <- extract_path("services$api_services$plot$client_plot")
path
}
