% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_methods.R
\name{get_property}
\alias{get_property}
\title{Extracts Property from ChatRBox S7 Objects}
\usage{
get_property(x, ...)
}
\arguments{
\item{x}{\code{S7} object. A \code{ChatRBox_obj} \code{S7} object. Required.}

\item{...}{Additional arguments including \code{property}, 
the property name to extract. Required.}
}
\value{
The requested property in the associated format.
}
\description{
This function returns given properties from the \code{ChatRBox_obj} \code{S7} object.
}
\examples{
# Defines tool function
add <- function(x,y) {
  x + y
}

# Generates S7 chatbot object with tool function
object <- object_generate(tools_list = list(add_two_numbers = add))

# Extracts tool environment from S7 object
tools <- get_property(object, "tools_env")

# Extracts services list from S7 object
services <- get_property(object, "services")

# These both contains the 'add' function under the name 'add_two_numbers'
ls(tools)
get_service_keys(services)

# This converts contents of tools into list for AI prompt
env_to_str(tools)

# Example AI response for addition of two numbers
response <- 'This service adds two numbers together!```json
{
 "path": "services$tools$add_two_numbers",
 "name": "add",
 "x": 70,
 "y": 76
}
```'
# Populates 'add' client function with extracted arguments from user question
result <- llm_api_result(llm_response = response,
                         object = object)

# Outputs tool result without accompanying AI dialogue
result

# Defines second tool function
subtract <- function(x,y) {
  x - y
}

# Outputs list of updated S7 properties given second tool function and 
# existing S7 tools environment
# Non-referenced arguments take default values
property_generate(object = object,
                  tools_env = object@tools_env,
                  tools_list = list(subtract_two_numbers = subtract))


# property_generate() is used as a helper function to update the S7 
# chatbot object in-place
# Here, the existing S7 tools environment is a default
update <- object_update(object = object,
                        tools_list = list(subtract_two_numbers = subtract))

# Extracts tool environment from updated S7 object
updated_tools <- get_property(update, "tools_env")

# This contains the 'add' and 'subtract' functions
ls(updated_tools)

# This converts contents of updated_tools into list for AI prompt
env_to_str(updated_tools)
}
