% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_message.R
\name{http_status_from_condition}
\alias{http_status_from_condition}
\title{Extracts HTTP Status Code from Error Conditions}
\usage{
http_status_from_condition(e)
}
\arguments{
\item{e}{Error condition. An HTTP request error condition to extract the status code from. Required.}
}
\value{
This function returns the extracted HTTP error status code as an integer. If unavailable, this function returns \code{NA_integer_}.
}
\description{
This function extracts the HTTP status code from failed API calls to present, alongside suggested next steps, in the ChatRBox dynamic error message.
}
\details{
The dynamic error message within \code{\link{llm_api_result}} outputs for any failed \code{do.call()}, populated with the attempted client path and parameters, API HTTP error, \pkg{httr2} configuration used by the client function and suggested next steps. This acknowledges that multiple components under various controls may contribute towards a failed \code{do.call()}. For example, if using external API services with generated client functions, failures may arise from access issues or timeouts. Whilst the chatbot prompt is precise, incorrect client paths or parameters may arise from AI hallucinations. Therefore, the dynamic error message presents users with maximum information such that they may trouble-shoot failed \code{do.call()} calls by adjusting what they deem the most necessary. This function extracts the API service HTTP error which is not only presented in this error message, but used to print the relevant suggested improvements using \code{\link{http_status_guidance}}. Lastly, this function was designed to be as robust as possible, whereby it attempts to extract error codes from multiple different structures like nested lists and from within character warnings. In the absence of valid extraction using these methods, this function returns \code{NA}.
}
\examples{
# A dynamic error message outputs for any failed AI-facilitated do.call()
# This contains attempted client path and parameters, API HTTP error, httr2 
# configuration used by the client function and suggested next steps  

# Example HTTP error status
error_condition <- structure(
  list(status_code = 414),
  class = "error"
)

# Extracts error code to assign ChatRBox-specific guidance
status_code <- http_status_from_condition(error_condition)
status_code

# Assigns ChatRBox guidance to extracted code
guidance <- http_status_guidance(status_code)

# Example external API error message
api_error <- "URL too long"

# Final error message depicts both general API warning and assigned 
# ChatRBox guidance
error <- add_http_status_guidance(error_condition, api_error) 
}
\seealso{
\code{\link{llm_api_result}}, \code{\link{http_status_guidance}}
}
