% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_message.R
\name{http_status_guidance}
\alias{http_status_guidance}
\title{Returns User Advice per HTTP Status Code}
\usage{
http_status_guidance(status)
}
\arguments{
\item{status}{Integer. Integer HTTP status code extracted using \code{\link{http_status_from_condition}}. Required.}
}
\value{
This function returns a character string specific to each HTTP status code, with advice on how users may resolve failed \code{do.call()} calls. Each advice block is presented below the HTTP status code within the dynamic error message and specific to ChatRBox workflows. If no status code is inputted, this function returns an empty string.
}
\description{
This function matches inputted HTTP integer status codes to string advice blocks. These strings suggest next steps for users to resolve failed \code{do.call()} calls, and are appended to the ChatRBox dynamic error message.
}
\examples{
# A dynamic error message outputs for any failed AI-facilitated do.call()
# This contains attempted client path and parameters, API HTTP error, httr2 
# configuration used by the client function and suggested next steps  

# Example HTTP error status
error_condition <- structure(
  list(status_code = 414),
  class = "error"
)

# Extracts error code to assign ChatRBox-specific guidance
status_code <- http_status_from_condition(error_condition)
status_code

# Assigns ChatRBox guidance to extracted code
guidance <- http_status_guidance(status_code)

# Example external API error message
api_error <- "URL too long"

# Final error message depicts both general API warning and assigned 
# ChatRBox guidance
error <- add_http_status_guidance(error_condition, api_error) 
}
\seealso{
\code{\link{http_status_guidance}}
}
