% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_installation.R
\name{internal_system.file}
\alias{internal_system.file}
\title{Locates ChatRBox File in inst/ Folder}
\usage{
internal_system.file(..., package = "ChatRBox", mustWork = FALSE)
}
\arguments{
\item{...}{Path within inst/. Required.}

\item{package}{Package name. Defaults to "ChatRBox".}

\item{mustWork}{Error if file not found. Defaults to FALSE.}
}
\value{
inst/ file path as character string.
}
\description{
This function is used to locate the system prompt Markdown files during ChatRBox package loading. These are converted into strings and used in \code{\link{ChatRBox}} and \code{\link{object_generate}} as the \code{prompt_template} argument. Hence, users may easily customize prompts as simple Markdown files as to control AI model behavior.
}
\examples{
file_path <- internal_system.file("prompt", paste0("prompt", ".md"), package = "ChatRBox")
file_path
}
