% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_installation.R
\name{load_prompt_md}
\alias{load_prompt_md}
\title{Loads Markdown File using Absolute Path}
\usage{
load_prompt_md(path)
}
\arguments{
\item{path}{Character. The path to the location of the Markdown file as a string; the file is presumed to contain an AI prompt. Hence, \code{path} must finish with the name of the desired Markdown file, including \code{.md}. Required.}
}
\value{
The text contained in the specified Markdown file as a character string. When \code{\link{load_prompt_md}} is used inside a \code{\link{ChatRBox}} chatbot object to load a prompt, this string will not only enable AI instruction but have \code{paths_string}, \code{args_string} and \code{past_outputs} interpolated into it, given that these names are present and encased within \code{{}}.
}
\description{
This function loads Markdown files from local environments via absolute paths. It is intended to be used as the \code{prompt_template} argument during chatbot initialization, setting \code{path} as the path to a Markdown file with an AI prompt.
}
\details{
\code{path} should not include the working directory as this is extracted and prepended to the inputted path using \code{normalizePath}. For example, the path to the default prompt in the ChatRBox package is \code{inst/prompt/prompt.md}, which would act as the \code{path} argument. This has no reference to working directories, usernames or project titles such that this function may be universally used. This function enables users to personalize their prompts with ease but it is advised that users append their desired instruction to the default prompt provided. The default prompt concerns the background process which is required for any API service or R tool function workflow.
}
