% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_installation.R
\name{load_prompt_template}
\alias{load_prompt_template}
\title{Loads ChatRBox Prompt Templates from inst/ Folder as String}
\usage{
load_prompt_template(prompt_name = "prompt", package = "ChatRBox")
}
\arguments{
\item{prompt_name}{Character. Name of the prompt file (without .md). Defaults to "prompt".}

\item{package}{Character. Package name. Defaults to "ChatRBox".}
}
\value{
Character string of the prompt template Markdown file.
}
\description{
This function loads the ChatRBox system prompt Markdown files in inst/prompt as character strings.
}
\details{
This function is used to define the \code{prompt_template} default values in \code{\link{ChatRBox}} and \code{\link{object_generate}}. This prompt includes reference to \code{paths_string}, \code{args_string} and \code{past_outputs}, which can be interpolated after \code{ChatRBox_obj} \code{S7} object generation due to string conversion. This function is used similarly for the \code{summary_list} argument in \code{\link{ChatRBox}}, whereby any service without an equivalently named summary prompt (assuming the \code{$talk()} argument \code{summarize} is set to \code{TRUE}) is assigned the default summary prompt also located in the inst folder. Hence, this enables users to provide extra summary information and specific AI-generated summary instructions per provided service.
}
\examples{
# Loads default prompt from inst/prompt
prompt <- load_prompt_template()
prompt
}
