% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_token.R
\name{make_req}
\alias{make_req}
\title{Constructs a httr2 API Request with Authentication}
\usage{
make_req(
  base_url = NULL,
  path = NULL,
  token = NULL,
  auth_scheme = NULL,
  token_header = NULL,
  timeout = 60,
  method = "GET",
  verbose = FALSE
)
}
\arguments{
\item{base_url}{Character. Base API URL (without trailing slash). Defaults to \code{NULL}.}

\item{path}{Character. The path to the relevant endpoint. Defaults to \code{NULL}.}

\item{token}{Character. Optional API authentication token. Defaults to \code{NULL}.}

\item{auth_scheme}{Character. API authorization scheme. Defaults to \code{NULL}, given \code{token_header} is populated.}

\item{token_header}{Character. API token header name if \code{auth_scheme} is \code{NULL}. Defaults to \code{NULL}.}

\item{timeout}{Numeric. Timeout value. Defaults to 60 seconds.}

\item{method}{Character. API HTTP method. Defaults to "GET" for likely chatbot API services.}

\item{verbose}{Logical. Whether to print verbose information. Defaults to FALSE.}
}
\value{
A \code{httr2_request} object representing an HTTP request, with the method, URL path, headers, timeout, and authentication applied as specified by the function arguments.
This request object can be executed with \code{httr2::req_perform()} to make the API call and is used within \code{\link{get_api_client_fns}} to extract external API client functions for chatbot access.
}
\description{
This function takes user HTTP request preferences to construct a \pkg{httr2} request body to facilitate API token usage.
}
\details{
This function acts as a helper within \code{\link{get_api_client_fns}} to extract client function R source code from the API endpoint \code{client_fns}. These have been separated to optimize API token handling, whereby tokens and authorization methods may be chosen to access protected APIs.
}
\examples{
# Constructs request body with fictitious inputs
req <- make_req(
 base_url = "https://api.example.com",
 path = "example/path",
 token = "FAKE_TOKEN",
 auth_scheme = "Bearer",
 token_header = "Authorization")
 
req
}
\seealso{
\code{\link{get_api_client_fns}}
}
