% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_api_client.R
\name{or_null}
\alias{or_null}
\title{Null-Coalescing Operator}
\usage{
or_null(a, b)
}
\arguments{
\item{a}{An object. The primary value to test for \code{NULL}. Required.}

\item{b}{An object. The value to return if \code{a} is \code{NULL}. Required.}
}
\value{
The value of \code{a} if it is not \code{NULL}; otherwise, the value of \code{b}.
}
\description{
Returns the first argument if it is not \code{NULL}, otherwise returns the second argument. Often used to provide a default value if the first is missing or \code{NULL}.
}
\details{
This function is used to provide default parameter values if none are supplied. This is used in \code{\link{build_api_client}} to create client functions for API services using the OpenAPI JSON schema.
}
\examples{
result1 <- or_null(NULL, "default")
result2 <- or_null(42, "default")
}
\seealso{
\code{\link{build_api_client}}
}
