% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_methods.R, R/methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{readPNG}
\alias{writePNG}
\alias{\%>\%}
\alias{.data}
\alias{.env}
\alias{eval_tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{png}{\code{\link[png]{readPNG}}, \code{\link[png]{writePNG}}}

  \item{rlang}{\code{\link[rlang:dot-data]{.data}}, \code{\link[rlang:dot-data]{.env}}, \code{\link[rlang]{eval_tidy}}}
}}

