% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_methods.R
\name{start_api_service}
\alias{start_api_service}
\title{Starts the Local Plotting API Service}
\usage{
start_api_service(port = 8000, host = "127.0.0.1")
}
\arguments{
\item{port}{Port. Port to launch API. Defaults to 8000}

\item{host}{Host. Host for API. Defaults to "127.0.0.1"}
}
\value{
A \pkg{callr} process handle. Use \code{$kill()} to stop the API service.
}
\description{
This function starts a local \pkg{plumber2} API service which may be provided to chatbots for linear plotting. This is mainly utilized for example and instruction scripts.
}
\details{
This function is used to locally start a linear plotting API service, to demonstrate the process by which users may supply AI chatbots with remote \pkg{plumber2} APIs. \pkg{plumber2} is the successor to \pkg{plumber} and adds asynchronous request handling. The service is firstly initiated using \code{proc <- ChatRBox::start_api_service()}, followed by \code{Sys.sleep(2)} to provide sufficient time and finally \code{proc$kill()} to disable the service. See \code{examples.R} in the examples folder and the ChatRBox vignette for further detail.
}
\examples{
# Simulates deployment of remote API service using the linear plotting API, 
# stored locally as plumber.R in inst/api_service 
proc <- start_api_service()

# Allows time for API service to launch in background R process
Sys.sleep(2) 

# Returns TRUE if process is created
proc$is_alive()

# Retrieves client functions from API URL (default host/port)
\dontrun{
  client_envir <- get_api_client_fns("http://127.0.0.1:8000")
  ls(client_envir)
}
}
