% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_methods.R
\name{store_keyring}
\alias{store_keyring}
\title{Stores Config File Values in Keyring}
\usage{
store_keyring(config_path = NULL, service = "ChatRBox")
}
\arguments{
\item{config_path}{File path. Path to YAML config file. Defaults to \code{NULL}. Users are advised to name their YAML config file "ChatRBox_config.yml", stored in their home directory.}

\item{service}{Character. Name of keyring service. Defaults to "ChatRBox".}
}
\value{
This function includes an invisible return as it stores all environment variables listed in the inputted config file into the keyring.
}
\description{
Reads a YAML config file and adds each entry to the keyring under the specified service.
}
\details{
\code{\link{start_api_service}} uses the \pkg{callr} function \code{r_bg}. This contains a background R process that temporarily copies user .Renviron files. Hence, users are advised to create a config file containing necessary environment variables (such as \pkg{ellmer} base URLs) and store these with \pkg{keyring}. This function facilitates this, whereby variables can be loaded using \code{\link{get_keyring}}. Note that user .Renviron files are only copied into the R session's temporary directory, such that \code{callr::r_bg()} is safe and adheres to file operation guidance.
}
\examples{
# Store config values with keyring, retrieve during initialization
\dontrun{
store_keyring(config_path = "~/ChatRBox_config.yml")
  chat <- ellmer::chat_ollama(base_url = get_keyring("OLLAMA_HOST"),
                              model    = get_keyring("OLLAMA_MODEL"))
}
}
